requirements.txt
Flask
requests
Flask-Session
werkzeug
gunicorn==21.2.0
pillow>=4.1.0
pilgram2
NumPy
opencv-python
realesrgan
super-image
image-enhancement